Move probabilty any worm tests into own file
